requirements.txt
africastalking==1.2.9
alembic==1.14.0
Authlib==1.3.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
cryptography==43.0.3
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
# -Flask-MySQLdb==2.0.0 # REMOVE THIS
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
# -mysqlclient==2.2.6 # REMOVE THIS
packaging==24.2
pluggy==1.5.0
# -pycparser==2.22 # Might be a dependency of Flask-SQLAlchemy, keep if `pip install -r` demands it
# -PyMySQL==1.1.1 # REMOVE THIS
psycopg2-binary==2.9.9 # ADD THIS
pytest==8.3.3
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3